Refactored EntryArray with a subclass LazyEntryArray where the serialized key value does not need less operator.
